+++ rbkApiModules.Commons.Core/Messaging/FileResult.cs
@@ -1,0 +1,8 @@
+namespace rbkApiModules.Commons.Core;
+
+public class FileResult
+{
+    public required MemoryStream FileStream { get; init; }
+    public required string FileName { get; init; }
+    public required string ContentType { get; init; }
+}